Avoid using AndroidJUnit4 runner to test Entrypoint implementations.
