+++ features/lockscreen/impl/src/test/kotlin/io/element/android/features/lockscreen/impl/DefaultLockScreenEntryPointTest.kt
@@ -7,19 +7,23 @@
 
 package io.element.android.features.lockscreen.impl
 
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.bumble.appyx.core.modality.BuildContext
+import com.bumble.appyx.testing.junit4.util.MainDispatcherRule
 import com.google.common.truth.Truth.assertThat
 import io.element.android.features.lockscreen.api.LockScreenEntryPoint
-import io.element.android.features.lockscreen.impl.unlock.activity.PinUnlockActivity
 import io.element.android.tests.testutils.lambda.lambdaError
 import io.element.android.tests.testutils.node.TestParentNode
+import org.junit.Rule
 import org.junit.Test
-import org.junit.runner.RunWith
 
-@RunWith(AndroidJUnit4::class)
 class DefaultLockScreenEntryPointTest {
+    @get:Rule
+    val instantTaskExecutorRule = InstantTaskExecutorRule()
+
+    @get:Rule
+    val mainDispatcherRule = MainDispatcherRule()
+
     @Test
     fun `test node builder Setup`() {
         val entryPoint = DefaultLockScreenEntryPoint()
@@ -60,12 +64,5 @@ class DefaultLockScreenEntryPointTest {
         assertThat(result).isInstanceOf(LockScreenFlowNode::class.java)
         assertThat(result.plugins).contains(LockScreenFlowNode.Inputs(LockScreenFlowNode.NavTarget.Settings))
         assertThat(result.plugins).contains(callback)
-    }
-
-    @Test
-    fun `test pin unlock intent`() {
-        val entryPoint = DefaultLockScreenEntryPoint()
-        val result = entryPoint.pinUnlockIntent(InstrumentationRegistry.getInstrumentation().context)
-        assertThat(result.component?.className).isEqualTo(PinUnlockActivity::class.qualifiedName)
     }
 }
